03_application-team/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.3
psycopg2==2.9.6
bcrypt==4.0.1
sshtunnel==0.4.0
requests
flask_caching

Werkzeug==3.0.0


tensorflow


transformers

numpy

matplotlib
pandas>=1.4.3

openpyxl


openai~=0.27.8

faiss-cpu~=1.7.4
gensim~=4.3.2

nltk~=3.8.1
torch-geometric~=2.4.0
datasets~=2.13.1

accelerate==0.31.0
bitsandbytes~=0.42.0

tqdm~=4.65.0



-f https://download.pytorch.org/whl/torch_stable.html
#torch>=2.0.1
torch==2.0.1+cpu



